コンソールのコマンドはすべて移植完了。 起動時の画面がかなりすっきりした。
git-svn-id: https://svn.osdn.jp/svnroot/chnosproject/beta@280 447e56a9-82e3-4d55-9dd4-5b2b0d47896b

+++ tolset_chn_000/chnos_009/chnos/io.c
@@ -1,8 +1,9 @@
 
 #include "core.h"
 
-uint IO_Read_CMOS(uchar addr)
+uchar IO_Read_CMOS(uchar addr)
 {
 	IO_Out8(0x70, addr);
 	return IO_In8(0x71);
 }
+
